+++ CHANGELOG.md
@@ -66,6 +66,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * The `branches`/`tags`/`git_refs`/`git_head` template keywords now return a
   list of `RefName`s. They were previously pre-formatted strings.
 
+* The new template keywords `local_branches`/`remote_branches` are added to show
+  only local/remote branches.
+
 ### Fixed bugs
 
 * Updating the working copy to a commit where a file that's currently ignored